### 2016-06-24 #31 @whedon
Submission: Prism: Multiple spline regression with regularization, dimensionality reduction, and feature selection
**Submitting author:** @cMadan (<a href="http://orcid.org/0000-0003-3228-6501">Christopher R Madan</a>)
**Repository:** <a href="https://github.com/cMadan/prism" target ="_blank">https://github.com/cMadan/prism</a>
**Version:** v2.0.0
**Archive:** <a href="http://dx.doi.org/10.5281/zenodo.56821" target ="_blank">http://dx.doi.org/10.5281/zenodo.56821</a>
