### 2016-06-26 fcc35b2
Updated code comment
