### 2016-06-26 b4f5eb6
Changed bib to just have first initials
