+++ paper/paper.bib
@@ -3,20 +3,20 @@
   year  = {1975},
   volume = {4},
   pages = {1--17},
-  author = {Grace Wahba and Svante Wold},
+  author = {G Wahba and S Wold},
   title = {A completely automatic french curve: Fitting spline functions by cross validation},
   journal = {Communications in Statistics}
 }
 @book{fox2000nonparametric,
   title = {Nonparametric simple regression: Smoothing scatterplots},
-  author = {John Fox},
+  author = {J Fox},
   year = {2000},
   publisher = {Sage}
 }
 @article{Tipp2000,  
   year = {2000},
   pages = {652--658},
-  author = {Michael E. Tipping},
+  author = {M E Tipping},
   title = {The relevance vector machine},
   journal = {Advances in Neural Information Processing Systems (NIPS)},
 }
@@ -24,13 +24,13 @@
   year = {2001},
   volume = {1},
   pages = {211--244},
-  author = {Michael E. Tipping},
+  author = {M E Tipping},
   title = {Sparse Bayesian learning and the relevance vector machine},
   journal = {Journal of Machine Learning Research},
 }
 @article{FaulTipp2003,  
   year = {2003},
-  author = {Michael E. Tipping and Anita C. Faul},
+  author = {M E Tipping and A C Faul},
   title = {Fast marginal likelihood maximisation for sparse Bayesian models},
   journal = {Proceedings of the Ninth International Workshop on Artificial Intelligence and Statistics (AISTATS)},
 }
@@ -39,7 +39,7 @@
   year  = {2003},
   volume = {6},
   pages = {309--315},
-  author = {Elizabeth R. Sowell and Bradley S. Peterson and Paul M. Thompson and Suzanne E. Welcome and Amy L. Henkenius and Arthur W. Toga},
+  author = {E R Sowell and B S Peterson and P M Thompson and S E Welcome and A L Henkenius and A W Toga},
   title = {Mapping cortical change across the human life span},
   journal = {Nature Neuroscience}
 }
@@ -48,7 +48,7 @@
   year  = {2013},
   volume = {23},
   pages = {2521--2530},
-  author = {L. J. Hogstrom and L. T. Westlye and K. B. Walhovd and A. M. Fjell},
+  author = {L J Hogstrom and L T Westlye and K B Walhovd and A M Fjell},
   title = {The structure of the cerebral cortex across adult life: Age-related patterns of surface area, thickness, and gyrification},
   journal = {Cerebral Cortex}
 }
@@ -57,7 +57,7 @@
   year  = {2016},
   volume = {134},
   pages = {617--629},
-  author = {Christopher R. Madan and Elizabeth A. Kensinger},
+  author = {C R Madan and E A Kensinger},
   title = {Cortical complexity as a measure of age-related brain atrophy},
   journal = {{NeuroImage}}
 }
@@ -66,7 +66,7 @@
   year  = {2010},
   volume = {50},
   pages = {1376--1383},
-  author = {Anders. M. Fjell and Kristine B. Walhovd and Lars T. Westlye and Ylva {\O}stby and Christian K. Tamnes and Terry L. Jernigan and Anthony Gamst and Anders M. Dale},
+  author = {A M Fjell and K B Walhovd and L T Westlye and Y {\O}stby and C K Tamnes and T L Jernigan and A Gamst and A M Dale},
   title = {When does brain aging accelerate? Dangers of quadratic fits in cross-sectional studies},
   journal = {{NeuroImage}}
 }
@@ -75,7 +75,7 @@
   year  = {2013},
   volume = {34},
   pages = {2239--2247},
[…]
