### 2016-06-26 #31 @cMadan
For the paper PDF, references after the first one aren't displayed quite right. Would using a semicolon between them (rather than a comma) be the fix?
