### 2016-06-27 #31 @arfon
> For the paper PDF, references after the first one aren't displayed quite right. Would using a semicolon between them (rather than a comma) be the fix?

@cMadan - yep, I've fixed this in https://github.com/cMadan/prism/pull/1 and updated the PDF in the review issue.
