### 2016-06-27 #31 @Kevin-Mattheus-Moerman
I have reviewed this submission in terms of MATLAB implementation and installation. I have also verified the demo and benchmark work as described. However I am not an expert in the technical topic the software addresses. Would it be possible to provide a more graphical display of the functionality and technical claims of the software, i.e. plot graphs and figures visualizing the nature of the input data, the processing and the output results? Perhaps if the functionality can be described more thoroughly and graphically in this manor users new to this field may be able to grasp functionality better.

I tested this submission using MATLAB R2016b PreRelease.

I can confirm installation (of the software and the third party SparseBayes V2 toolbox) and implementation proceeded as described in the README.md.

I was able to install the software and run the demo.m.

Comments on paper:
Check spelling for “lregularization” in paper
Fix spelling error “which camn be obtained” in paper
Does the figure need a caption/description, like in the README.md?
Check the references. Is the author missing for the reference: “Sparse Bayesian Learning and the Relevance Vector Machine”

In README.md check spelling for: “regression lead to”. Should this be “leads”?
