### 2016-06-27 #31 @labarba
This opens a bigger question for me—as JOSS editor—, which is: Should we be publishing packages that rely on proprietary programming and execution environments?

A quick look at the README for this submission immediately struck me with: _"requires three first-party toolboxes: (1) Curve Fitting Toolbox; (2) Statistics and Machine Learning Toolbox; (3) Signal Processing Toolbox"_
Are these paid MATLAB Toolboxes?

Should we have a conversation about that on another Issue or in the Google Group?
