### 2016-06-27 #31 @arfon
> This opens a bigger question for me—as JOSS editor—, which is: Should we be publishing packages that rely on proprietary programming and execution environments?

Yes I think these should be acceptable with the caveat that it should to be possible for a reviewer to address the following review points:

> [ ] Installation: Does installation proceed as outlined in the documentation?
> [ ] Functionality: Have the functional claims of the software been confirmed?
> [ ] Performance: Have any performance claims of the software been confirmed?

> Should we have a conversation about that on another Issue or in the Google Group?

Let's have this conversation in an issue. Are you happy to start it?
