### 2016-06-27 1e0cf4a
Edits from reviewer comments
