+++ paper/paper.md
@@ -12,7 +12,7 @@ authors:
  - name: Christopher R Madan
    orcid: 0000-0003-3228-6501
    affiliation: Boston College
-date: 26 June 2016
+date: 27 June 2016
 bibliography: paper.bib
 ---
 
@@ -22,7 +22,8 @@ Prism uses a combination of statistical methods to conduct spline-based multiple
 
 Prism has been tested in MATLAB 2015b and requires three first-party toolboxes: (1) Curve Fitting Toolbox; (2) Statistics and Machine Learning Toolbox; (3) Signal Processing Toolbox. Relevance vector regression requires the SparseBayes V2 toolbox, which can be obtained from http://www.relevancevector.com.
 
--![Illustration of Prism regression procedure, first conducting spline regression for each predictor, followed by dimensionality reduction and feature selection (panel A). The logo for Prism is shown in panel B.](fig1_prism.png)
+-![Illustration of Prism regression procedure.](fig1_prism.png)
+Figure: Illustration of Prism regression procedure, first conducting spline regression for each predictor, followed by dimensionality reduction and feature selection (panel A). The logo for Prism is shown in panel B.
 
 Prism was implemented for conducting multiple regression investigating age-related differences in brain morphology. While linear and quadratic regression are often used (e.g., [@SoweEtal2003; @HogsEtal2013; @MadaKens2016]), it has been shown that non-linear (spline) regression is more appropriate [@FjelEtal2010; @FjelEtal2013].
 
