### 2016-06-27 #31 @cMadan
@Kevin-Mattheus-Moerman, thanks for the review! I've fixed the typos and missing figure caption (was in paper.md, but not PDF). Reference is correct (is same as previous, 'line' notation is produced when PDF is compiled).
In demo.m, I have now added some scatter plots for evaluating the predictions.

@arfon, the subsequent references still look incorrect in the PDF. I've also made a few changes, so maybe you could re-compile the PDF?

@labarba, yes, these are paid MATLAB toolboxes, though (most) educational institutions that have MATLAB licences would also have these toolboxes. More broadly, I think any restriction on dependencies  would preclude many MATLAB-based projects from being published in JOSS.
