### 2016-06-28 40f1b2e
Fixing bibtex entries
