+++ paper/paper.bib
@@ -7,7 +7,7 @@
   title = {A completely automatic french curve: Fitting spline functions by cross validation},
   journal = {Communications in Statistics}
 }
-@book{fox2000nonparametric,
+@book{Fox2000,
   title = {Nonparametric simple regression: Smoothing scatterplots},
   author = {J Fox},
   year = {2000},
@@ -28,7 +28,7 @@
   title = {Sparse Bayesian learning and the relevance vector machine},
   journal = {Journal of Machine Learning Research},
 }
-@article{FaulTipp2003,  
+@article{TippFaul2003,  
   year = {2003},
   author = {M E Tipping and A C Faul},
   title = {Fast marginal likelihood maximisation for sparse Bayesian models},
