### 2016-06-28 #31 @arfon
@cMadan - I've recompiled the paper [here](https://github.com/openjournals/joss-reviews/files/336251/10.21105.joss.00031.pdf). Could you take a look and see if it's formatted better now?
