### 2016-06-28 #31 @cMadan
@arfon - much better! The only remaining issue is this "(e.g., (Sowell et al. 2003; Hogstrom et al. 2013; Madan and Kensinger 2016))". I'm not sure how I should change the md to make that turn out better. Effectively want something equivalent to "\cite<e.g.,>{REFS}".
