### 2016-06-28 #31 @nicoguaro
I think that instead of the logo an image of a result would be much illustrative.
