### 2016-06-28 4e1bea0
Added example output figure.
