### 2016-06-28 #31 @arfon
@cMadan - changing the Markdown to:

```
...quadratic regression are often used (e.g., @SoweEtal2003; @HogsEtal2013; @MadaKens2016), it has been shown that...
```

results in this output:

![screen shot 2016-06-27 at 10 00 11 pm](https://cloud.githubusercontent.com/assets/4483/16403504/25401396-3cb3-11e6-83fb-2ee17ff6aff5.png)

Does that look better? I'm not a pandoc expert I'm afraid.
