### 2016-06-28 3b9ca53
Fixing refs in paper.md
