+++ paper/paper.md
@@ -25,7 +25,7 @@ Prism has been tested in MATLAB 2015b and requires three first-party toolboxes: 
 -![Illustration of Prism regression procedure.](fig1_prism.png)
 Figure 1: Illustration of Prism regression procedure, first conducting spline regression for each predictor, followed by dimensionality reduction and feature selection (panel A). The logo for Prism is shown in panel B.
 
-Prism was implemented for conducting multiple regression investigating age-related differences in brain morphology. While linear and quadratic regression are often used (e.g., [@SoweEtal2003; @HogsEtal2013; @MadaKens2016]), it has been shown that non-linear (spline) regression is more appropriate [@FjelEtal2010; @FjelEtal2013].
+Prism was implemented for conducting multiple regression investigating age-related differences in brain morphology. While linear and quadratic regression are often used (e.g., @SoweEtal2003; @HogsEtal2013; @MadaKens2016), it has been shown that non-linear (spline) regression is more appropriate [@FjelEtal2010; @FjelEtal2013].
 
 # Example implementation
 
