### 2016-06-28 #31 @cMadan
@nicoguaro - I've just added example output.

@arfon - that's sufficiently better, thanks! I really appreciate the help! I've updated paper.md on the repo.
