### 2016-06-28 #31 @arfon
> @arfon - that's sufficiently better, thanks! I really appreciate the help! I've updated paper.md on the repo.

OK great. Here's the updated paper: [10.21105.joss.00031.pdf](https://github.com/openjournals/joss-reviews/files/337245/10.21105.joss.00031.pdf)
