### 2016-06-28 3429184
Fixed typo in image markdown code
