### 2016-06-28 342b74f
Changed how prediction error is calculated in output figure
