### 2016-06-30 #31 @Kevin-Mattheus-Moerman
Thanks for implementing the changes and for adding basic visualizations of results in demo.m. For the paper it seems that the reference with the title: "Sparse Bayesian Learning and the Relevance Vector Machine" misses an author field. Should this be Michael E. Tipping?
I've now ticked all the boxes above. Once the reference for the paper is verified I can recommend this paper for publication in JOSS.
