### 2016-06-30 #31 @Kevin-Mattheus-Moerman
There is as you know an ongoing discussion on the inclusion of software relying on proprietary programming environments. It may be that a final decision on your submission will be made after that discussion has been concluded. Apologies for the delay.
