### 2016-06-30 #31 @nicoguaro
Even though I'm not an editor I think that the output of the discussion should not influence the publication of this paper.  This paper was submitted before the discussion even started, if it would have started before maybe the author would have not submitted.
