### 2016-06-30 #31 @cMadan
@Kevin-Mattheus-Moerman - The author is Tipping, but it is being displayed 'properly'. The PDF is compiled from markdown+bib by @arfon's code, and is based on an IEEE convention. See here: http://tex.stackexchange.com/questions/29381/is-it-normal-for-bibtex-to-replace-similar-author-names-with
