### 2016-07-01 #31 @arfon
OK thanks for the review here @Kevin-Mattheus-Moerman!

@cMadan - this is now accepted into JOSS, your DOI is: http://dx.doi.org/10.21105/joss.00031

🎉 🚀 💥
