### 2016-07-01 #31 @cMadan
I just noticed that under the second figure, it says "Figure 1" (that isn't in the markdown itself). I made a change to paper.md that hopefully fixes this.
