### 2016-07-02 #31 @arfon
@cMadan - could you take a look now and if this is better? https://github.com/openjournals/joss-papers/blob/master/joss.00031/10.21105.joss.00031.pdf
